requirements.txt
torch
pandas
matplotlib
scikit-learn
ncps
rich
pyyaml
hydra-core